requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
sqlite3
requests==2.25.1 
pandas==1.3.0 
numpy==1.21.0